Push the "first user" logic into the model layer, add tests for this
